Display original request URI in NoResourceFoundException message
This commit ensures that the original request URI is displayed in
`NoResourceFoundException` error messages when logged. Without this
change, it can be confusing to see only the attempted resource path.
There are cases where the original request was not meant for resource
handling and we want to understand why this wasn't processed by another
handler.

The Problem Detail attribute has not been changed as the "instance"
attribute already displays the request path.

Closes gh-34553