Add `ImportHttpServices.GroupProvider` support
Update `@ImportHttpServices` to allow the group to either be specified
using the `group` attribute or a `groupProvider` implementation.

Using a `groupProvider` allows the group name to be deduced from the
class metadata. For example, users could use a custom annotation
to hold the group name.

Groups must either be specified using the `group` attribute *or* the
`groupProvider` attribute. Attempting to set both will cause an
`IllegalStateException` to be thrown.

See gh-35447