Fix spliterator characteristics in ConcurrentReferenceHashMap
The Spliterators returned by values, entrySet, and keySet incorrectly reported
the SIZED characteristic, instead of CONCURRENT. This could lead to bugs when
the map is concurrently modified during a stream operation.

For keySet and values, the incorrect characteristics are inherited from
AbstractMap, so to rectify that the respective methods are overridden and
custom collections are provided that report the correct Spliterator
characteristics.

Signed-off-by: Patrick Strawderman <pstrawderman@netflix.com>